track player setup
