+++ react-native.config.js
@@ -6,4 +6,10 @@ module.exports = {
       },
     },
   },
+
+  project: {
+    ios: {},
+    android: {},
+  },
+  assets: ['./assets/fonts/'],
 };
