+++ src/components/search/filterSearch.tsx
@@ -1,10 +1,11 @@
-import { SurahItem } from '../../constants/surahList.ts';
+import { SurahItem } from '../../constants/surahName/surahList.ts';
 
 const filterSurah = (data: SurahItem[], searchStr = '') => {
   if (searchStr !== '') {
     return data.filter(
-      ({ title, arabicTittle }) => title.toLowerCase().includes(searchStr.toLowerCase())
-        || arabicTittle.toLowerCase().includes(searchStr.toLowerCase()),
+      ({ title, arabicTittle }) =>
+        title.toLowerCase().includes(searchStr.toLowerCase()) ||
+        arabicTittle.toLowerCase().includes(searchStr.toLowerCase())
     );
   }
   return data;
