--- src/constants/highlightingData/highlightingData.ts
@@ -1,182 +1,0 @@
-const highlightingData = {
-  a_rahman_sudais: {
-    1: [
-      {
-        id: '01_01_01',
-        verseNumber: 1,
-        startTime: 0,
-        endTime: 1,
-        word: 'ٱلْحَمْدُ',
-      },
-      {
-        id: '01_01_02',
-        verseNumber: 1,
-        startTime: 1,
-        endTime: 2,
-        word: 'لِلَّهِ',
-      },
-      {
-        id: '01_01_03',
-        verseNumber: 1,
-        startTime: 2,
-        endTime: 3,
-        word: 'رَبِّ',
-      },
-      {
-        id: '01_01_04',
-        verseNumber: 1,
-        startTime: 3,
-        endTime: 4,
-        word: 'ٱلْعَٰلَمِينَ ﴿١﴾ ‏',
-      },
-      {
-        id: '01_02_01',
-        verseNumber: 2,
-        startTime: 5,
-        endTime: 6,
-        word: 'ٱلرَّحْمَٰنِ',
-      },
-      {
-        id: '01_02_02',
-        verseNumber: 2,
-        startTime: 6,
-        endTime: 7.2,
-        word: 'ٱلرَّحِيمِ ﴿٢﴾‏',
-      },
-      {
-        id: '01_03_01',
-        verseNumber: 3,
-        startTime: 7.5,
-        endTime: 8.5,
-        word: 'مَـٰلِكِ',
-      },
-      {
-        id: '01_03_02',
-        verseNumber: 3,
-        startTime: 8.6,
-        endTime: 9.4,
-        word: 'يَوْمِ',
-      },
-      {
-        id: '01_03_03',
-        verseNumber: 3,
-        startTime: 9.7,
-        endTime: 10.5,
-        word: 'ٱلدِّينِ ﴿٣﴾‏',
-      },
-      {
-        id: '01_04_01',
-        verseNumber: 4,
-        startTime: 11.1,
-        endTime: 11.7,
-        word: 'إِيَّاكَ',
-      },
-      {
-        id: '01_04_02',
-        verseNumber: 4,
-        startTime: 11.8,
-        endTime: 12.7,
-        word: 'نَعْبُدُ',
-      },
-      {
-        id: '01_04_03',
-        verseNumber: 4,
-        startTime: 12.8,
-        endTime: 13.7,
-        word: 'وَإِيَّاكَ',
-      },
-      {
-        id: '01_04_04',
-        verseNumber: 4,
-        startTime: 13.8,
-        endTime: 15.5,
-        word: 'نَسْتَعِينُ ﴿٤﴾‏',
-      },
-      {
-        id: '01_05_01',
-        verseNumber: 5,
-        startTime: 15.5,
-        endTime: 16.7,
-        word: 'ٱهْدِنَا',
-      },
-      {
-        id: '01_05_02',
-        verseNumber: 5,
-        startTime: 16.8,
-        endTime: 18,
-        word: 'ٱلصِّرَٰطَ',
-      },
-      {
-        id: '01_05_03',
-        verseNumber: 5,
-        startTime: 18.1,
-        endTime: 19.5,
-        word: 'ٱلْمُسْتَقِيمَ ﴿٥﴾‏',
-      },
-      {
-        id: '01_06_01',
-        verseNumber: 6,
-        startTime: 20,
-        endTime: 21,
-        word: 'صِرَٰطَ',
-      },
-      {
-        id: '01_06_02',
-        verseNumber: 6,
-        startTime: 21.2,
-        endTime: 22.1,
-        word: 'ٱلَّذِينَ',
-      },
-      {
-        id: '01_06_03',
-        verseNumber: 6,
-        startTime: 22.2,
-        endTime: 22.9,
-        word: 'أَنْعَمْتَ',
-      },
-      {
-        id: '01_06_04',
-        verseNumber: 6,
-        startTime: 23,
-        endTime: 23.9,
-        word: 'عَلَيْهِمْ',
-      },
-      {
-        id: '01_06_05',
-        verseNumber: 6,
-        startTime: 24,
-        endTime: 24.9,
-        word: 'غَيْرِ',
-      },
-      {
-        id: '01_06_06',
-        verseNumber: 6,
-        startTime: 25,
-        endTime: 26,
-        word: 'ٱلْمَغْضُوبِ',
-      },
-      {
-        id: '01_06_07',
-        verseNumber: 6,
-        startTime: 26,
-        endTime: 27,
-        word: 'عَلَيْهِمْ',
-      },
-      {
-        id: '01_06_08',
-        verseNumber: 6,
-        startTime: 27,
-        endTime: 28,
-        word: 'وَلَا',
-      },
-      {
-        id: '01_06_09',
-        verseNumber: 6,
-        startTime: 28,
-        endTime: 32,
-        word: 'ٱلضَّآلِّينَ ﴿٦﴾‏',
-      },
-    ],
-  },
-};
-export default highlightingData;
