--- src/constants/reciterData.ts
@@ -1,12 +1,0 @@
-const recitersData = [
-  {
-    id: 1,
-    name: 'Abdul Rehman Al Sudais',
-  },
-  {
-    id: 2,
-    name: 'Ahmed Al Ajmy',
-  },
-];
-
-export default recitersData;
