--- src/constants/surahList.ts
@@ -1,695 +1,0 @@
-export type SurahItem = {
-  title: string;
-  arabicTittle: string;
-  englishTitle: string;
-  surahNo: number;
-};
-
-const surahList: Array<SurahItem> = [
-  {
-    surahNo: 1,
-    title: 'Al Fatiha',
-    arabicTittle: 'الفَاتِحَة',
-    englishTitle: 'The Opening',
-  },
-  {
-    surahNo: 2,
-    title: 'Al Baqarah',
-    arabicTittle: 'البَقَرَة',
-    englishTitle: 'The Cow',
-  },
-  {
-    surahNo: 3,
-    title: 'Al Imran',
-    arabicTittle: 'آلِ عِمْرَان',
-    englishTitle: 'The Family of Imran',
-  },
-  {
-    surahNo: 4,
-    title: 'An Nisa',
-    arabicTittle: 'النِّسَاء',
-    englishTitle: 'The Women',
-  },
-  {
-    surahNo: 5,
-    title: 'Al Maidah',
-    arabicTittle: 'المَائِدَة',
-    englishTitle: 'The Table Spread',
-  },
-  {
-    surahNo: 6,
-    title: 'Al Anam',
-    arabicTittle: 'الأَنْعَام',
-    englishTitle: 'The Cattle',
-  },
-  {
-    surahNo: 7,
-    title: 'Al Araf',
-    arabicTittle: 'الأَعْرَاف',
-    englishTitle: 'The Heights',
-  },
-  {
-    surahNo: 8,
-    title: 'Al Anfal',
-    arabicTittle: 'الأَنْفَال',
-    englishTitle: 'The Spoils of War',
-  },
-  {
-    surahNo: 9,
-    title: 'At Tawbah',
-    arabicTittle: 'التَّوْبَة',
-    englishTitle: 'The Repentance',
-  },
-  {
-    surahNo: 10,
-    title: 'Yunus',
-    arabicTittle: 'يُونُس',
-    englishTitle: 'Jonah',
-  },
-  {
-    surahNo: 11,
-    title: 'Hud',
-    arabicTittle: 'هُود',
-    englishTitle: 'Prophet Hud',
-  },
-  {
-    surahNo: 12,
-    title: 'Yusuf',
-    arabicTittle: 'يُوسُف',
-    englishTitle: 'Joseph',
-  },
-  {
-    surahNo: 13,
-    title: 'Ar Rad',
-    arabicTittle: 'الرَّعْد',
-    englishTitle: 'The Thunder',
-  },
-  {
-    surahNo: 14,
-    title: 'Ibrahim',
-    arabicTittle: 'إِبْرَاهِيم',
-    englishTitle: 'Abraham',
-  },
-  {
-    surahNo: 15,
-    title: 'Al Hijr',
-    arabicTittle: 'الحِجْر',
-    englishTitle: 'The Rocky Tract',
-  },
-  {
-    surahNo: 16,
-    title: 'An Nahl',
-    arabicTittle: 'النَّحْل',
-    englishTitle: 'The Bee',
-  },
-  {
-    surahNo: 17,
-    title: 'Al Isra',
-    arabicTittle: 'الإِسْرَاء',
-    englishTitle: 'The Night Journey',
-  },
-  {
-    surahNo: 18,
-    title: 'Al Kahf',
-    arabicTittle: 'الكَهْف',
-    englishTitle: 'The Cave',
-  },
-  {
-    surahNo: 19,
-    title: 'Maryam',
-    arabicTittle: 'مَرْيَم',
-    englishTitle: 'Maryam',
-  },
-  {
-    surahNo: 20,
-    title: 'Ta Ha',
-    arabicTittle: 'طٰهٰ',
-    englishTitle: 'Ta Ha',
-  },
-  {
-    surahNo: 21,
-    title: 'Al Anbiya',
-    arabicTittle: 'الأَنْبِيَاء',
-    englishTitle: 'The Prophets',
-  },
-  {
-    surahNo: 22,
-    title: 'Al Hajj',
-    arabicTittle: 'الحَجّ',
-    englishTitle: 'The Pilgrimage',
-  },
-  {
-    surahNo: 23,
-    title: 'Al Muminun',
-    arabicTittle: 'المُؤْمِنُون',
-    englishTitle: 'The Believers',
-  },
-  {
-    surahNo: 24,
-    title: 'An Nur',
-    arabicTittle: 'النُّور',
-    englishTitle: 'The Light',
-  },
-  {
-    surahNo: 25,
-    title: 'Al Furqan',
-    arabicTittle: 'الفُرْقَان',
-    englishTitle: 'The Criterion',
-  },
-  {
-    surahNo: 26,
-    title: 'Ash Shuara',
-    arabicTittle: 'الشُّعَرَاء',
-    englishTitle: 'The Poets',
-  },
-  {
-    surahNo: 27,
-    title: 'An Naml',
-    arabicTittle: 'النَّمْل',
-    englishTitle: 'The Ant',
-  },
-  {
-    surahNo: 28,
-    title: 'Al Qasas',
-    arabicTittle: 'القَصَص',
-    englishTitle: 'The Stories',
-  },
-  {
-    surahNo: 29,
-    title: 'Al Ankabut',
-    arabicTittle: 'العَنْكَبُوت',
-    englishTitle: 'The Spider',
-  },
-  {
-    surahNo: 30,
-    title: 'Ar Rum',
-    arabicTittle: 'الرُّوم',
-    englishTitle: 'The Romans',
-  },
-  {
-    surahNo: 31,
-    title: 'Luqman',
-    arabicTittle: 'لُقْمَان',
-    englishTitle: 'Luqman',
-  },
-  {
-    surahNo: 32,
-    title: 'As Sajdah',
-    arabicTittle: 'السَّجْدَة',
-    englishTitle: 'The Prostration',
-  },
-  {
-    surahNo: 33,
-    title: 'Al Ahzab',
-    arabicTittle: 'الأَحْزَاب',
-    englishTitle: 'The Clans',
-  },
-  {
-    surahNo: 34,
-    title: 'Saba',
-    arabicTittle: 'سَبَأ',
-    englishTitle: 'Sheba',
-  },
-  {
-    surahNo: 35,
-    title: 'Fatir',
-    arabicTittle: 'فاطر',
-    englishTitle: 'The Originator',
-  },
-  {
-    surahNo: 36,
-    title: 'Ya Sin',
-    arabicTittle: 'يٰس',
-    englishTitle: 'Ya Sin',
-  },
-  {
-    surahNo: 37,
-    title: 'As Saffat',
-    arabicTittle: 'الصَّافَّات',
-    englishTitle: 'Those who set the Ranks',
-  },
-  {
-    surahNo: 38,
-    title: 'Sad',
-    arabicTittle: 'ص',
-    englishTitle: 'Sad',
-  },
-  {
-    surahNo: 39,
-    title: 'Az Zumar',
-    arabicTittle: 'الزُّمَر',
-    englishTitle: 'The Groups',
-  },
-  {
-    surahNo: 40,
-    title: 'Ghafir',
-    arabicTittle: 'غَافِر',
-    englishTitle: 'The Forgiver',
-  },
-  {
-    surahNo: 41,
-    title: 'Fussilat',
-    arabicTittle: 'فُصِّلَت',
-    englishTitle: 'Explained in Detail',
-  },
-  {
-    surahNo: 42,
-    title: 'Ash Shura',
-    arabicTittle: 'الشُّورىٰ',
-    englishTitle: 'The Consultation',
-  },
-  {
-    surahNo: 43,
-    title: 'Az Zukhruf',
-    arabicTittle: 'الزُّخْرُف',
-    englishTitle: 'Ornaments of Gold',
-  },
-  {
-    surahNo: 44,
-    title: 'Ad Dukhan',
-    arabicTittle: 'الدُّخَان',
-    englishTitle: 'The Smoke',
-  },
-  {
-    surahNo: 45,
-    title: 'Al Jathiyah',
-    arabicTittle: 'الجَاثِيَة',
-    englishTitle: 'The Crouching',
-  },
-  {
-    surahNo: 46,
-    title: 'Al Ahqaf',
-    arabicTittle: 'الأَحْقَاف',
-    englishTitle: 'The Wind Curved Sandhills',
-  },
-  {
-    surahNo: 47,
-    title: 'Muhammad',
-    arabicTittle: 'مُحَمَّد',
-    englishTitle: 'Muhammad',
-  },
-  {
-    surahNo: 48,
-    title: 'Al Fath',
-    arabicTittle: 'الفَتْح',
-    englishTitle: 'The Victory',
-  },
-  {
-    surahNo: 49,
-    title: 'Al Hujurat',
-    arabicTittle: 'الحُجُرَات',
-    englishTitle: 'The Rooms',
-  },
-  {
-    surahNo: 50,
-    title: 'Qaf',
-    arabicTittle: 'ق',
-    englishTitle: 'Qaf',
-  },
-  {
-    surahNo: 51,
-    title: 'Az Zariyat',
-    arabicTittle: 'الذَّارِيَات',
-    englishTitle: 'The Winnowing Winds',
-  },
-  {
-    surahNo: 52,
-    title: 'At Tur',
-    arabicTittle: 'الطُّور',
-    englishTitle: 'The Mount',
-  },
-  {
-    surahNo: 53,
-    title: 'An Najm',
-    arabicTittle: 'النَّجْم',
-    englishTitle: 'The Star',
-  },
-  {
-    surahNo: 54,
-    title: 'Al Qamar',
-    arabicTittle: 'القَمَر',
-    englishTitle: 'The Moon',
-  },
-  {
-    surahNo: 55,
-    title: 'Ar Rahman',
-    arabicTittle: 'الرَّحْمَٰن',
-    englishTitle: 'The Beneficent',
-  },
-  {
-    surahNo: 56,
-    title: 'Al Waqiah',
-    arabicTittle: 'الوَاقِعَة',
-    englishTitle: 'The Event',
-  },
-  {
-    surahNo: 57,
-    title: 'Al Hadid',
-    arabicTittle: 'الحَدِيد',
-    englishTitle: 'The Iron',
-  },
-  {
-    surahNo: 58,
-    title: 'Al Mujadila',
-    arabicTittle: 'المُجَادِلَة',
-    englishTitle: 'The Pleading Woman',
-  },
-  {
-    surahNo: 59,
-    title: 'Al Hashr',
-    arabicTittle: 'الحَشْر',
-    englishTitle: 'The Gathering',
-  },
-  {
-    surahNo: 60,
-    title: 'Al Mumtahinah',
-    arabicTittle: 'المُمْتَحَنَة',
-    englishTitle: 'She that is to be examined',
-  },
-  {
-    surahNo: 61,
-    title: 'As Saff',
-    arabicTittle: 'الصَّفّ',
-    englishTitle: 'The Ranks',
-  },
-  {
-    surahNo: 62,
-    title: 'Al Jumuah',
-    arabicTittle: 'الجُمُعَة',
-    englishTitle: 'Friday',
-  },
-  {
-    surahNo: 63,
-    title: 'Al Munafiqun',
-    arabicTittle: 'المُنَافِقُون',
-    englishTitle: 'The Hypocrites',
-  },
-  {
-    surahNo: 64,
-    title: 'At Taghabun',
-    arabicTittle: 'التَّغَابُن',
-    englishTitle: 'Mutual Disillusion',
-  },
-  {
-    surahNo: 65,
-    title: 'At Talaq',
-    arabicTittle: 'الطَّلَاق',
-    englishTitle: 'Divorce',
-  },
-  {
-    surahNo: 66,
-    title: 'At Tahrim',
-    arabicTittle: 'التَّحْرِيم',
-    englishTitle: 'The Prohibition',
-  },
-  {
-    surahNo: 67,
-    title: 'Al Mulk',
-    arabicTittle: 'المُلْك',
-    englishTitle: 'The Sovereignty',
-  },
-  {
-    surahNo: 68,
-    title: 'Al Qalam',
-    arabicTittle: 'القَلَم',
-    englishTitle: 'The Pen',
-  },
-  {
-    surahNo: 69,
-    title: 'Al Haqqah',
-    arabicTittle: 'الحَاقَّة',
-    englishTitle: 'The Reality',
-  },
-  {
-    surahNo: 70,
-    title: 'Al Maarij',
-    arabicTittle: 'المَعَارِج',
-    englishTitle: 'The Ascending Stairways',
-  },
-  {
-    surahNo: 71,
-    title: 'Nuh',
-    arabicTittle: 'نُوح',
-    englishTitle: 'Noah',
-  },
-  {
-    surahNo: 72,
-    title: 'Al Jinn',
-    arabicTittle: 'الجِنّ',
-    englishTitle: 'The Jinn',
-  },
-  {
-    surahNo: 73,
-    title: 'Al Muzzammil',
-    arabicTittle: 'المُزَّمِّل',
-    englishTitle: 'The Enshrouded One',
-  },
-  {
-    surahNo: 74,
-    title: 'Al Muddathir',
-    arabicTittle: 'المُدَّثِّر',
-    englishTitle: 'The Cloaked One',
-  },
-  {
-    surahNo: 75,
-    title: 'Al Qiyamah',
-    arabicTittle: 'القِيَامَة',
-    englishTitle: 'The Resurrection',
-  },
-  {
-    surahNo: 76,
-    title: 'Al Insan',
-    arabicTittle: 'الإِنْسَان',
-    englishTitle: 'Man',
-  },
-  {
-    surahNo: 77,
-    title: 'Al Mursalat',
-    arabicTittle: 'المُرْسَلَات',
-    englishTitle: 'The Emissaries',
-  },
-  {
-    surahNo: 78,
-    title: 'An Naba',
-    arabicTittle: 'النَّبَأ',
-    englishTitle: 'The Tidings',
-  },
-  {
-    surahNo: 79,
-    title: 'An Naziat',
-    arabicTittle: 'النَّازِعَات',
-    englishTitle: 'Those who drag forth',
-  },
-  {
-    surahNo: 80,
-    title: 'Abasa',
-    arabicTittle: 'عَبَسَ',
-    englishTitle: 'He frowned',
-  },
-  {
-    surahNo: 81,
-    title: 'At Takwir',
-    arabicTittle: 'التَّكْوِير',
-    englishTitle: 'The Overthrowing',
-  },
-  {
-    surahNo: 82,
-    title: 'Al Infitar',
-    arabicTittle: 'الإِنْفِطَار',
-    englishTitle: 'The Cleaving',
-  },
-  {
-    surahNo: 83,
-    title: 'Al Mutaffifin',
-    arabicTittle: 'المُطَفِّفِين',
-    englishTitle: 'Defrauding',
-  },
-  {
-    surahNo: 84,
-    title: 'Al Inshiqaq',
-    arabicTittle: 'الإِنْشِقَاق',
-    englishTitle: 'The Splitting Open',
-  },
-  {
-    surahNo: 85,
-    title: 'Al Buruj',
-    arabicTittle: 'البُرُوج',
-    englishTitle: 'The Mansions of the Stars',
-  },
-  {
-    surahNo: 86,
-    title: 'At Tariq',
-    arabicTittle: 'الطَّارِق',
-    englishTitle: 'The Morning Star',
-  },
-  {
-    surahNo: 87,
-    title: 'Al Ala',
-    arabicTittle: 'الأَعْلَىٰ',
-    englishTitle: 'The Most High',
-  },
-  {
-    surahNo: 88,
-    title: 'Al Ghashiyah',
-    arabicTittle: 'الغَاشِيَة',
-    englishTitle: 'The Overwhelming',
-  },
-  {
-    surahNo: 89,
-    title: 'Al Fajr',
-    arabicTittle: 'الفَجْر',
-    englishTitle: 'The Dawn',
-  },
-  {
-    surahNo: 90,
-    title: 'Al Balad',
-    arabicTittle: 'البلد',
-    englishTitle: 'The City',
-  },
-  {
-    surahNo: 91,
-    title: 'Ash Shams',
-    arabicTittle: 'الشَّمْس',
-    englishTitle: 'The Sun',
-  },
-  {
-    surahNo: 92,
-    title: 'Al Lail',
-    arabicTittle: 'اللَّيْل',
-    englishTitle: 'The Night',
-  },
-  {
-    surahNo: 93,
-    title: 'Ad Duha',
-    arabicTittle: 'الضُّحَىٰ',
-    englishTitle: 'The Morning Hours',
-  },
-  {
-    surahNo: 94,
-    title: 'Al Inshirah',
-    arabicTittle: 'الشَّرْح',
-    englishTitle: 'The Relief',
-  },
-  {
-    surahNo: 95,
-    title: 'At Tin',
-    arabicTittle: 'التِّين',
-    englishTitle: 'The Fig',
-  },
-  {
-    surahNo: 96,
-    title: 'Al Alaq',
-    arabicTittle: 'العَلَق',
-    englishTitle: 'The Clot',
-  },
-  {
-    surahNo: 97,
-    title: 'Al Qadr',
-    arabicTittle: 'القَدْر',
-    englishTitle: 'The Power',
-  },
-  {
-    surahNo: 98,
-    title: 'Al Bayyinah',
-    arabicTittle: 'البَيِّنَة',
-    englishTitle: 'The Clear Proof',
-  },
-  {
-    surahNo: 99,
-    title: 'Az Zalzalah',
-    arabicTittle: 'الزَّلْزَلَة',
-    englishTitle: 'The Earthquake',
-  },
-  {
-    surahNo: 100,
-    title: 'Al Adiyat',
-    arabicTittle: 'العَادِيَات',
-    englishTitle: 'The Courser',
-  },
-  {
-    surahNo: 101,
-    title: 'Al Qariah',
-    arabicTittle: 'القَارِعَة',
-    englishTitle: 'The Calamity',
-  },
-  {
-    surahNo: 102,
-    title: 'At Takathur',
-    arabicTittle: 'التَّكَاثُر',
-    englishTitle: 'The Rivalry in world increase',
-  },
-  {
-    surahNo: 103,
-    title: 'Al Asr',
-    arabicTittle: 'العَصْر',
-    englishTitle: 'The Declining Day',
-  },
-  {
-    surahNo: 104,
-    title: 'Al Humazah',
-    arabicTittle: 'الهُمَزَة',
-    englishTitle: 'The Traducer',
-  },
-  {
-    surahNo: 105,
-    title: 'Al Fil',
-    arabicTittle: 'الفِيل',
-    englishTitle: 'The Elephant',
-  },
-  {
-    surahNo: 106,
-    title: 'Quraish',
-    arabicTittle: 'قُرَيْش',
-    englishTitle: 'Quraysh',
-  },
-  {
-    surahNo: 107,
-    title: 'Al Maun',
-    arabicTittle: 'المَاعُون',
-    englishTitle: 'The Small Kindness',
-  },
-  {
-    surahNo: 108,
-    title: 'Al kausar',
-    arabicTittle: 'الكَوْثَر',
-    englishTitle: 'Abundance',
-  },
-  {
-    surahNo: 109,
-    title: 'Al Kafirun',
-    arabicTittle: 'الكَافِرُون',
-    englishTitle: 'The Disbelievers',
-  },
-  {
-    surahNo: 110,
-    title: 'An Nasr',
-    arabicTittle: 'النَّصْر',
-    englishTitle: 'The Help',
-  },
-  {
-    surahNo: 111,
-    title: 'Al lahab',
-    arabicTittle: 'اللهب',
-    englishTitle: 'The Palm Fibre',
-  },
-  {
-    surahNo: 112,
-    title: 'Al Ikhlas',
-    arabicTittle: 'الإِخْلَاص',
-    englishTitle: 'The Sincerity',
-  },
-  {
-    surahNo: 113,
-    title: 'Al Falaq',
-    arabicTittle: 'الفَلَق',
-    englishTitle: 'The Daybreak',
-  },
-  {
-    surahNo: 114,
-    title: 'An Nas',
-    arabicTittle: 'النَّاس',
-    englishTitle: 'The Mankind',
-  },
-];
-
-export default surahList;
