+++ src/constants/surahText/surah_01.ts
@@ -1,0 +1,44 @@
+const surah01 = [
+  {
+    id: '01_00',
+    verseNumber: 0,
+    ayah: 'أَعُوْذُ بِاللّٰهِ مِنَ الشَّيْطٰانِ الرَّجِيْمِ',
+  },
+  {
+    id: '02_01',
+    verseNumber: 1,
+    ayah: 'بِسْمِ اللَّهِ الرَّحْمَٰنِ الرَّحِيمِ ﴿١﴾',
+  },
+  {
+    id: '02_02',
+    verseNumber: 2,
+    ayah: 'الْحَمْدُ لِلَّهِ رَبِّ الْعَالَمِينَ ﴿٢﴾',
+  },
+  {
+    id: '02_03',
+    verseNumber: 3,
+    ayah: 'الرَّحْمَٰنِ الرَّحِيمِ ﴿٣﴾',
+  },
+  {
+    id: '02_04',
+    verseNumber: 4,
+    ayah: 'مَالِكِ يَوْمِ الدِّينِ ﴿٤﴾',
+  },
+  {
+    id: '02_05',
+    verseNumber: 5,
+    ayah: 'إِيَّاكَ نَعْبُدُ وَإِيَّاكَ نَسْتَعِينُ ﴿٥﴾',
+  },
+  {
+    id: '02_06',
+    verseNumber: 6,
+    ayah: 'اهْدِنَا الصِّرَاطَ الْمُسْتَقِيمَ ﴿٦﴾',
+  },
+  {
+    id: '02_07',
+    verseNumber: 7,
+    ayah: 'صِرَاطَ الَّذِينَ أَنْعَمْتَ عَلَيْهِمْ غَيْرِ الْمَغْضُوبِ عَلَيْهِمْ وَلَا الضَّالِّينَ ﴿٧﴾',
+  },
+];
+
+export default surah01;
