+++ src/index.d.ts
@@ -1,1 +1,2 @@
 declare module '*.png';
+declare module '*.jpeg';
