+++ src/screens/audio/audioScreen.tsx
@@ -1,80 +1,99 @@
 import Slider from '@react-native-community/slider';
-import { useNavigation } from '@react-navigation/native';
-import React, { useEffect, useState } from 'react';
+import { useNavigation, useRoute } from '@react-navigation/native';
+import React, {
+  useCallback, useEffect, useMemo, useState,
+} from 'react';
 import {
-  SafeAreaView, Text, View, ImageBackground,
+  SafeAreaView, Text, View, ImageBackground, ScrollView,
 } from 'react-native';
 import TrackPlayer from 'react-native-track-player';
 import AntDesign from 'react-native-vector-icons/AntDesign';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 
-import BackGroundImage from '../../assets/images/borderImage/border1.png';
-import highlightingData from '../../constants/highlightingData/highlightingData.ts';
+import BackGroundImage from '../../assets/images/borderImage/border.png';
+import audioPath from '../../constants/audioData/audioDataPath.ts';
+import getSurahText from '../../constants/codeSplitting/mappingSurahData.ts';
 import useTheme from '../../theming/useTheme.ts';
 import useThemedStyles from '../../theming/useThemedStyles.ts';
-import MediaPlayer from '../../utils/MediaPlayer.ts';
+import { setupPlayer } from '../../utils/trackPlayerServices';
 
 import styles from './style.ts';
 
-const mediaPlayer = MediaPlayer.getInstance();
-
 function AudioScreen() {
   const navigation = useNavigation();
-
+  const route = useRoute();
+  const { title, surahNo } = route.params;
   const style = useThemedStyles(styles);
   const theme = useTheme();
   const [isPlay, setIsPlay] = useState(false);
   const [currentPosition, setCurrentPosition] = useState(0);
   const [duration, setDuration] = useState(0);
   const [currentVerse, setCurrentVerse] = useState(1);
+  const verses = useMemo(() => audioPath[surahNo - 1]?.verses || [], [surahNo]);
+  const ayahNo = audioPath[surahNo - 1]?.verses[currentVerse - 1]?.ayahNo;
 
-  const handleBackPress = () => {
+  const handleBackPress = useCallback(() => {
     navigation.navigate('Home' as never);
+    TrackPlayer.reset();
+  }, [navigation]);
+
+  // Memoized function to set up the player
+  const initPlayer = useCallback(async () => {
+    await TrackPlayer.reset();
+    await TrackPlayer.add(verses);
+  }, [verses]);
+
+  const setupAndAddPlayer = async () => {
+    await setupPlayer().then(async () => {
+      await initPlayer();
+    });
   };
 
+  // Effect to run the player setup
   useEffect(() => {
-    async function setup() {
-      await mediaPlayer.setup();
+    setupAndAddPlayer();
+  }, []);
 
-      const queue = await TrackPlayer.getQueue();
-      if (mediaPlayer.isSetup && queue.length <= 0) {
-        await mediaPlayer.addTracks();
+  useEffect(() => {
+    let isMounted = true;
+
+    const updatePosition = async () => {
+      const newPosition = await TrackPlayer.getPosition();
+      const newDuration = await TrackPlayer.getDuration();
+      if (isMounted) {
+        setCurrentPosition(newPosition);
+        setDuration(newDuration);
+        setTimeout(updatePosition, 1500);
       }
-    }
+    };
 
-    setup();
-  }, []);
+    updatePosition();
 
-  useEffect(() => {
-    const updatePosition = setInterval(async () => {
-      const {
-        currentPosition: playerPosition,
-        currentDuration,
-      } = await mediaPlayer.getPlayerState();
-      setCurrentPosition(playerPosition);
-      setDuration(currentDuration);
-    }, 1000);
+    return () => {
+      isMounted = false;
+    };
+  }, [currentPosition, currentVerse]);
 
-    return () => clearInterval(updatePosition);
-  }, []);
-
-  const handlePlayPress = async () => {
+  const handlePlayPress = useCallback(async () => {
+    // Check if the player is initialized before calling play/pause
     if (isPlay) {
-      await mediaPlayer.pause();
+      await TrackPlayer.pause();
     } else {
-      await mediaPlayer.play();
+      await TrackPlayer.play();
     }
-    setIsPlay(!isPlay);
-  };
+    setIsPlay((prevState) => !prevState);
+  }, [isPlay]);
 
-  const handleSkipBackward = async () => {
+  const handleSkipBackward = useCallback(async () => {
+    // Check if the player is initialized before skipping backward
     await TrackPlayer.skipToPrevious();
-  };
+  }, []);
 
-  const handleSkipForward = async () => {
-    await mediaPlayer.skipToNext();
-  };
+  const handleSkipForward = useCallback(async () => {
+    // Check if the player is initialized before skipping forward
+    await TrackPlayer.skipToNext();
+  }, []);
 
   const formatTime = (seconds: number) => {
     const minutes = Math.floor(seconds / 60);
@@ -84,9 +103,8 @@ function AudioScreen() {
 
   const getCurrentVerse = async () => {
     const currentTrack = await TrackPlayer.getCurrentTrack();
-    const currentData = highlightingData.a_rahman_sudais[1].find(
-      (data) => data.verseNumber === currentTrack,
-    );
+    const currentData = getSurahText(surahNo)?.find((data) => data.verseNumber === currentTrack);
+
     if (currentData) {
       setCurrentVerse(currentData.verseNumber + 1);
     }
@@ -94,7 +112,7 @@ function AudioScreen() {
 
   useEffect(() => {
     getCurrentVerse();
-  }, [currentPosition]);
+  });
 
   return (
     <SafeAreaView style={style.outerContainer}>
@@ -105,35 +123,28 @@ function AudioScreen() {
           color={theme.colors.SEARCHBAR}
           onPress={handleBackPress}
         />
-        <Text style={style.headerText}>Surah Fatiha</Text>
+        <Text style={style.headerText}>{title}</Text>
         <View />
       </View>
       <ImageBackground source={BackGroundImage} style={style.backgroundImage}>
         <View style={style.bannerView}>
-          <View style={style.textContainer}>
-            <Text style={style.surah}>
-              {highlightingData.a_rahman_sudais[1].map((wordObj, index) => {
-                const isPlayingWord = currentPosition >= wordObj.startTime
-                && currentPosition <= wordObj.endTime;
-                const isCurrentAyah = currentVerse === wordObj.verseNumber;
-
-                return (
+          <View>
+            <ScrollView>
+              <Text style={style.surah}>
+                {getSurahText(surahNo)?.map((wordObj) => (
                   <React.Fragment key={wordObj.id}>
-                    <Text
-                      style={[
-                        isPlayingWord && isCurrentAyah
-                        && style.highlightedWord,
-                        isCurrentAyah && !isPlayingWord
-                        && style.highlightedVerse,
-                      ]}
-                    >
-                      {wordObj.word}
-                    </Text>
-                    {index < highlightingData.a_rahman_sudais[1].length - 1 && <Text> </Text>}
+                    {ayahNo === wordObj.verseNumber && (
+                      // for ayah highlighting add this in following text
+                      // (style={ayahNo === wordObj.verseNumber ? style.highlightedVerse : null})
+                      <Text>
+                        {wordObj.ayah}
+                        {' '}
+                      </Text>
+                    )}
                   </React.Fragment>
-                );
-              })}
-            </Text>
+                ))}
+              </Text>
+            </ScrollView>
           </View>
         </View>
       </ImageBackground>
@@ -144,10 +155,10 @@ function AudioScreen() {
           value={currentPosition}
           minimumValue={0}
           maximumValue={duration}
-          onValueChange={(value) => mediaPlayer.seekTo(value)}
+          onValueChange={(value) => TrackPlayer.seekTo(value)}
           thumbTintColor={theme.colors.SEARCHBAR}
           minimumTrackTintColor={theme.colors.SEARCHBAR}
-          maximumTrackTintColor={theme.colors.SEARCHBARCONTAINER}
+          maximumTrackTintColor={theme.colors.TEXT_SECONDARY}
         />
         <View style={style.timerContainer}>
           <Text style={style.timerStyle}>{formatTime(currentPosition)}</Text>
