+++ src/screens/home/homeScreen.tsx
@@ -1,9 +1,7 @@
 import React, { useState } from 'react';
-import {
-  ImageBackground, SafeAreaView, StatusBar, StatusBarStyle,
-} from 'react-native';
+import { ImageBackground, SafeAreaView, StatusBar, StatusBarStyle } from 'react-native';
 
-import BackGroundImage from '../../assets/images/coverImages/launchScreen.png';
+import BackGroundImage from '../../assets/images/coverImages/launchScreen.jpeg';
 import SearchBar from '../../components/search/searchBar.tsx';
 import useTheme from '../../theming/useTheme.ts';
 import useThemedStyles from '../../theming/useThemedStyles.ts';
