+++ src/utils/trackPlayerServices.js
@@ -1,0 +1,58 @@
+import TrackPlayer, {
+  AppKilledPlaybackBehavior,
+  Capability,
+  Event,
+} from 'react-native-track-player';
+
+let isSetup = false;
+
+export async function setupPlayer() {
+  if (isSetup) {
+    return true; // Already set up, no need to initialize again
+  }
+
+  try {
+    await TrackPlayer.setupPlayer(); // Initialize TrackPlayer
+    await TrackPlayer.updateOptions({
+      // Set options for TrackPlayer
+      android: {
+        appKilledPlaybackBehavior: AppKilledPlaybackBehavior.StopPlaybackAndRemoveNotification,
+      },
+      capabilities: [
+        Capability.Play,
+        Capability.Pause,
+        Capability.SkipToNext,
+        Capability.SkipToPrevious,
+        Capability.SeekTo,
+      ],
+      compactCapabilities: [Capability.Play, Capability.Pause, Capability.SkipToNext],
+      progressUpdateEventInterval: 2,
+    });
+
+    isSetup = true;
+  } catch (error) {
+    isSetup = false;
+    console.log('Error setting up TrackPlayer:', error);
+    throw error;
+  }
+
+  return isSetup;
+}
+
+export async function playbackService() {
+  TrackPlayer.addEventListener(Event.RemotePause, () => {
+    TrackPlayer.pause();
+  });
+
+  TrackPlayer.addEventListener(Event.RemotePlay, () => {
+    TrackPlayer.play();
+  });
+
+  TrackPlayer.addEventListener(Event.RemoteNext, () => {
+    TrackPlayer.skipToNext();
+  });
+
+  TrackPlayer.addEventListener(Event.RemotePrevious, () => {
+    TrackPlayer.skipToPrevious();
+  });
+}
